Move room randomization into a separate class
